finished apolloserver

+++ graphqlServer/src/typeDefs.js
@@ -3,11 +3,11 @@ const { gql } = require('apollo-server')
 module.exports = gql`
 
   type ExposureFactors {
-    val3: Int!
-    val5: Int!
+    val3: Float!
+    val5: Float!
   }
 
   type Query {
-    exposure(personId: Int!): ExposureFactors!
+    exposure(personId: Float!): ExposureFactors!
   }
 `
